refactor: piece maker 패키지를 piece 하위 패키지로 이동
