+++ src/main/java/chess/domain/piece/maker/StartingPiecesGenerator.java
@@ -1,4 +1,4 @@
-package chess.domain.board.maker;
+package chess.domain.piece.maker;
 
 import chess.domain.Color;
 import chess.domain.File;
